pyproject.toml
[project]
name = "ai-ecom-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "imblearn>=0.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "langchain-text-splitters>=0.3.11",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "numpy>=2.3.3",
    "pypdf>=6.1.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "textblob>=0.19.0",
    "xgboost>=3.0.5",
]
